Created grid generator (#5)

+++ Finite Element Method - Runtime/src/reader/TXTFileReader.h
@@ -28,7 +28,7 @@ private:
         { GlobalDataEnum::DENSITY, [&](const std::string& value) { m_GlobalData.SetDensity(std::stod(value)); }},
         { GlobalDataEnum::SPECIFIC_HEAT, [&](const std::string& value) { m_GlobalData.SetSpecificHeat(std::stod(value)); }},
         { GlobalDataEnum::ELEMENTS_NUMBER, [&](const std::string& value) { m_GlobalData.ReserveGridElements(std::stoi(value)); }},
-        { GlobalDataEnum::NODES_NUMBER, [&](const std::string& value) { m_GlobalData.ReserveGridNodes(std::stod(value)); }}
+        { GlobalDataEnum::NODES_NUMBER, [&](const std::string& value) { m_GlobalData.ReserveGridNodes(std::stoi(value)); }}
     };
 };
